venv-requirements.txt
wheel==0.37.1
asgiref==3.4.1
Django==4.0.1
setuptools==44.0.0
sqlparse==0.4.2
mysqlclient==2.1.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
celery==5.2.6
